backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.3
pydantic-settings>=2.1.0
sqlalchemy>=2.0.25
duckdb>=1.0.0
pandas>=2.2.0
numpy>=1.26.3
scikit-learn>=1.4.0
python-multipart>=0.0.6
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
httpx>=0.26.0
aiofiles>=23.2.1
apscheduler>=3.10.4
plotly>=5.18.0
scipy>=1.12.0
